Add another small dataset for testing
Add a small sample of the Human Origins (HO) dataset to test the new
phenotype prediction functionality as well as the dummy phenotype
generator. The latter added two files: a .phe file with generated
phenotypes and a .png file visualizing them.

The sample is in the EIGENSTRAT format only, since it's the only format
accepted by the dummy phenotype generator for now.